Karger's Min Cut algo
This implemtation seems fairly quick, however it could definitely be
refactored a bit to be more readable. Notably, the handling of the node
indexes.. comsuming the file with base-1 nodes and putting them into an
array that is base-0 results in some confusing -1 and +1 ops. Also, this
currently ONLY returns the number of edges for the min cut, which isn't
all that helpful..where is the min cut? Which nodes are on which side??
Would be more useful to keep track of this as well.

Lastly, I've hardcoded 100 iterations of Karger's. For a 200 node graph
it is suggested to do 100k iterations, but I got a single failure in 20
runs with 100 so that just seems way overkill. And takes a long time to
run. (2 secs for 100 so probably ~2000 seconds or ~30 minutes for 100k)